Added

+++ database/migrations/2024_06_20_133059_add_foreign_to_blogs_table.php
@@ -11,26 +11,27 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('blogs', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('blog_category_id');
-            $table->string('title');
-            $table->string('slug');
-            $table->longText('description');
-            $table->longText('content');
-            $table->json('tags')->nullable();
-            $table->dateTime('publish_date')->nullable();
-            $table->string('featured_photo');
-            $table->string('file_id')->nullable();
-            $table->enum('status', ['active', 'inactive']);
+        if(!Schema::hasTable('blogs')){
+            Schema::create('blogs', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('blog_category_id');
+                $table->string('title');
+                $table->string('slug');
+                $table->longText('description');
+                $table->longText('content');
+                $table->json('tags')->nullable();
+                $table->dateTime('publish_date')->nullable();
+                $table->string('featured_photo');
+                $table->string('file_id')->nullable();
+                $table->enum('status', ['active', 'inactive']);
 
-            $table->index(['blog_category_id', 'slug']);
+                $table->index(['blog_category_id', 'slug']);
 
-            $table->foreign('blog_category_id')->references('id')->on('blog_categories')->onDelete('cascade');
+                $table->foreign('blog_category_id')->references('id')->on('blog_categories')->onDelete('cascade');
 
-            $table->timestamps();
-
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
